some minor updates (cmake, type_traits)

+++ src/main.cpp
@@ -5,9 +5,9 @@
 int main()
 {
     om::Device dv("cuda:0");
-    om::Tensor<float16_t> tf1({10, 10}, dv);
+    om::Tensor<float> tf1({10, 10}, dv);
     tf1.fill(10.0f);
 
-    om::Tensor<float16_t> res = tf1 * float16_t(2.0f);
+    om::Tensor<float> res = tf1 * float(2.0f);
     om::print(res);
 }
